Removed mention of letters as an image from puzzle description
There are a lot of directions that one can look for text in an image.
In my original writeup I gave away the fact that the letters were shown
in the image itself (and hence were not just ascii hidden somewhere in
the file).
I didn't want to give that hint, so it's now removed.